+++ example/lib/camera_page.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mrz_scanner/flutter_mrz_scanner.dart';
 
-import 'camera_overlay.dart';
-
 class CameraPage extends StatefulWidget {
   @override
   _CameraPageState createState() => _CameraPageState();
@@ -10,34 +8,19 @@ class CameraPage extends StatefulWidget {
 
 class _CameraPageState extends State<CameraPage> {
   bool isParsed = false;
-  MRZController controller;
+  MRZController? controller;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: const Text('Camera'),
-        ),
-        body: Column(
-          children: <Widget>[
-            Flexible(
-              flex: 3,
-              child: CameraOverlay(
-                child: MRZScanner(
-                  onControllerCreated: (controller) =>
-                      onControllerCreated(controller),
-                ),
-              ),
-            ),
-            Flexible(
-              child: Container(
-                  color: Theme.of(context).backgroundColor,
-                  child: const Center(
-                    child: Text('Flutter view below the native camera'),
-                  )),
-            ),
-          ],
-        ));
+      appBar: AppBar(
+        title: const Text('Camera'),
+      ),
+      body: MRZScanner(
+        withOverlay: true,
+        onControllerCreated: onControllerCreated,
+      ),
+    );
   }
 
   @override
@@ -71,7 +54,7 @@ class _CameraPageState extends State<CameraPage> {
                   Text('Expriy date: ${result.expiryDate}'),
                   Text('Personal number: ${result.personalNumber}'),
                   Text('Personal number 2: ${result.personalNumber2}'),
-                  RaisedButton(
+                  ElevatedButton(
                     child: const Text('ok'),
                     onPressed: () {
                       isParsed = false;